# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryEmailsByStatus } from './dashboardSummaryEmailsByStatus';
import type { DashboardSummaryProspectsByStatus } from './dashboardSummaryProspectsByStatus';

export interface DashboardSummary {
  totalProspects: number;
  totalEmailsSent: number;
  replyRate: number;
  totalCampaigns: number;
  emailsDraftedToday: number;
  /** @nullable */
  topCampaign: string | null;
  prospectsByStatus?: DashboardSummaryProspectsByStatus;
  emailsByStatus?: DashboardSummaryEmailsByStatus;
}