# lib/api-zod/src/generated/types/emailGenerateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Sales Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailGenerateInputTone } from './emailGenerateInputTone';

export interface EmailGenerateInput {
  prospectId: number;
  /** @nullable */
  campaignId?: number | null;
  tone?: EmailGenerateInputTone;
  context?: string;
}